Fix OSD_ListViewModes trying to return array on the stack
In C, returning an array returns a pointer to the first element, which will be on the stack in this case. As there's no guarantee that local data on the stack survives after returning from a function, this is dangerous.

Making the variable static to make it survive past the function lifetime.